move to abstract class
