+++ src/app/space/forge-wizard/forge-quickstart.component.html
@@ -1,10 +1,10 @@
 <pfng-wizard #wizard
              [config]="config">
 
-  <div [formGroup]="form" [ngSwitch]="history.currentGui.metadata?.name">
+  <div [formGroup]="form" [ngSwitch]="currentGui.metadata?.name">
     <pfng-wizard-step [config]="steps[0]">
       <span *ngSwitchCase="'Quickstart'">
-        <quickstart [gui]="history.currentGui" [form]="form"></quickstart>
+        <quickstart [gui]="currentGui" [form]="form"></quickstart>
       </span>
     </pfng-wizard-step>
   </div>
