+++ src/app/space/forge-wizard/forge-wizard.component.html
@@ -14,13 +14,13 @@
   </div>
 
   <div [formGroup]="form" [ngSwitch]="currentGui.metadata?.name">
-    <pfng-wizard-step [config]="stepCode">
-      <pfng-wizard-substep [config]="stepGithubImportPickOrganisation">
+    <pfng-wizard-step [config]="steps[0]">
+      <pfng-wizard-substep [config]="steps[3]">
       <span *ngSwitchCase="'io.fabric8.forge.generator.github.GithubImportPickOrganisationStep'">
         <organisation-step [gui]="currentGui" [form]="form"></organisation-step>
       </span>
       </pfng-wizard-substep>
-      <pfng-wizard-substep [config]="stepGithubRepositories">
+      <pfng-wizard-substep [config]="steps[4]">
       <span *ngSwitchCase="'io.fabric8.forge.generator.github.GithubImportPickRepositoriesStep'">
         <repositories-step [gui]= "currentGui" [form]="form"></repositories-step>
       </span>
@@ -28,22 +28,22 @@
     </pfng-wizard-step>
 
 
-    <pfng-wizard-step [config]="stepDeployment">
-      <pfng-wizard-substep [config]="stepConfigurePipeline">
+    <pfng-wizard-step [config]="steps[1]">
+      <pfng-wizard-substep [config]="steps[5]">
       <span *ngSwitchCase="'Obsidian: Configure Pipeline'">
         <pipeline-step [gui]="currentGui" [form]="form" [labelSpace]="currentSpace.attributes.name"></pipeline-step>
       </span>
       </pfng-wizard-substep>
 
-      <pfng-wizard-substep [config]="stepCreateBuildConfig">
+      <pfng-wizard-substep [config]="steps[6]">
       <span *ngSwitchCase="'io.fabric8.forge.generator.kubernetes.CreateBuildConfigStep'">
         <build-config-step [gui]="currentGui" [form]="form"></build-config-step>
       </span>
       </pfng-wizard-substep>
     </pfng-wizard-step>
 
-    <pfng-wizard-step [config]="stepReview">
-      <pfng-wizard-substep [config]="stepReviewConfig">
+    <pfng-wizard-step [config]="steps[2]">
+      <pfng-wizard-substep [config]="steps[7]">
       <span *ngSwitchCase="'Review'">
         <review-step [gui]="currentGui"></review-step>
       </span>
